requirements.txt
PyGithub>=2.1.1
schedule>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.2
rich>=13.7.0
click>=8.1.7
aiohttp>=3.9.1
jinja2>=3.1.2
pytest>=7.4.3
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1
